Review: moved code and improved some types
